+++ CHANGELOG.md
@@ -72,6 +72,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - No more jittering while running down slopes with the glider out
 - Axe normal attack rewards energy without skill points
 - Gliders no longer suffer from unreasonable amounts of induced drag
+- Camera is now clipping a lot less
 
 ## [0.10.0] - 2021-06-12